feat: improve ExecutionResult model (+ cascade changes)
