+++ components/clarity-repl/Cargo.toml
@@ -37,7 +37,7 @@ getrandom = { version = "0.2.3", features = ["js"] }
 rand_pcg = "0.3.1"
 rand_seeder = "0.2.2"
 atty = "0.2.14"
-clarity = { git = "https://github.com/stacks-network/stacks-blockchain.git", rev = "4ddffa6da4289d14f528a7af8e5369d25290e9e7", default-features = false, optional = true }
+clarity = { git = "https://github.com/stacks-network/stacks-blockchain.git", rev = "e7482540e17099698ab79aa428c9d9bf2e4ac281", default-features = false, optional = true }
 # clarity = { package = "clarity", path = "../../../stacks-blockchain/clarity", default-features = false, optional = true }
 
 # DAP Debugger
